+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Possibility to transfer claims via the sdk interface
 - Save states when switching orientation
 - Enabled app2app workflow
-- Fixed a bug with the soft flow if only one app was installed
+- Enabled handling of redirect uri
 
 ### Changed
 - Renamed NetID to netID
@@ -27,4 +27,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug with the soft flow if only one app was installed
 - Fixed a problem with finding the activities of a package
 - Fixed wrong scopes for hard/soft flow
+- Fixed a bug with the soft flow if only one app was installed